Rename errors file to index.js
